feat: 홈페이지 랜딩으로 수정
