+++ src/app/(home)/Page.module.css
@@ -1,6 +1,6 @@
 .hero__grid {
   background-size: cover;
-  background-image: url('../../../assets/image/grid.png');
+  background-image: url('../..//assets/image/grid.png');
 }
 
 .hero__container__wrap {
