--- src/app/(home)/_components/MainPage.module.css
@@ -1,39 +1,0 @@
-.container {
-  padding: 40px 0;
-}
-
-.category__container {
-  display: flex;
-  padding: 12px 16px;
-  margin: 24px 0 16px 0;
-  align-self: stretch;
-  border-radius: var(--border-radius-md);
-  background: var(--color-bg-tertiary);
-}
-
-.keyword__wrapper {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.keyword__container {
-  border-radius: var(--border-radius-md);
-  background: var(--black);
-  display: flex;
-  padding: 8px 16px;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 64px;
-}
-
-.keyword__container > span {
-  color: var(--white);
-}
-
-.button__wrapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 48px !important;
-}
